Revamp langsync to make much fewer API requests when writing
Instead of making an edit on every page no matter what, now only writes
a page if its text changes. This should drastically reduce the number
of requests as typically only a handful of pages need to be written. A
side effect is that the other pages are no longer null-edited, which
should be an acceptable tradeoff for much lower total script completion
times.