Fix edge-case error in langsync
Creating a new module page with wikitext content (e.g. a module
documentation subpage) throws a Lua error because the expected content
model is "Scribunto". Will now retry to create the page while forcing
the content model to "wikitext", but only if the original content model
from EN is also "wikitext".

Trying to keep-it-stupid-simple, so this retry has specific restrictions
and should only happen in this one particular case. Otherwise it might
introduce other, new errors.